Add files via upload
In the TotalBudgetCalculation system, a BudgetOptimizer class is created to facilitate budget calculations. This class allows users to input their own values or work with default values.

The calculateBudget method is used to calculate the budget by applying a specific formula, which takes xTarget as an input parameter. The value of xTarget is iteratively adjusted in the goalSeek function.

The goalSeek function initializes with a high value set to zApproved and a low value set to 0. It then calculates the xTarget value as the midpoint of the high and low values. This xTarget value is passed to the calculateBudget method to obtain a zTarget value.

The difference between zTarget and zApproved is then compared. If the difference is less than the STEP_SIZE of 0.1, the xTarget value is returned. If the difference is greater, the high value is updated to xTarget, and if the difference is lower, the low value is updated to xTarget.

Finally, the output is compared to determine if it is positive or zero. If it is, the xTarget value is displayed; otherwise, an error is reported.